backend/requirements.txt
numpy==1.24.4
pandas==2.1.4
uvicorn==0.24.0
fastapi==0.104.1
python-multipart==0.0.6
pytesseract==0.3.13
PyMuPDF==1.23.7
Pillow==10.0.0
pydantic>=2.6.0,<3.0.0
neo4j==5.14.1
langchain==0.0.350
spacy>=3.7.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
docling>=2.28.4
docling-core==2.36.0
docling-ibm-models==3.4.4
docling-parse==4.0.0
